% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjacent2edge.R
\name{adjacent2edge}
\alias{adjacent2edge}
\title{Function to convert tree in adjacent list format to edge
representation.}
\usage{
adjacent2edge(tree)
}
\arguments{
\item{tree}{The graph or tree in adjacent list format.}
}
\description{
Function to convert tree in adjacent list format to edge
representation.
}

