% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkObservationFlag.R
\name{checkObservationFlag}
\alias{checkObservationFlag}
\title{Check observation flag}
\usage{
checkObservationFlag(flag)
}
\arguments{
\item{flag}{A character (or character vector) of flag values to test for
validity.}
}
\value{
A logical vector of the same length as flag indicating if the flags
are valid (TRUE = valid).
}
\description{
Observation flags in the SWS are only allowed to take certain values.  This
function ensures that the supplied flags are all valid.
}

