% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeRatio.R
\name{computeRatio}
\alias{computeRatio}
\title{Function for computing the yield}
\usage{
computeRatio(numerator, denominator)
}
\arguments{
\item{numerator}{The value of the numerator.}

\item{denominator}{The value of the denominator.}
}
\description{
This is a function to compute yield based on production and area
harvested.
}

