% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faoswsUtil-package.R
\docType{package}
\name{faoswsUtil-package}
\alias{faoswsUtil-package}
\title{A suite of function to support the manipulation of data for
working with the FAO new statistical working system.}
\description{
This package provides extra functionality to the faosws package
for the manipulation of data.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

