% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeZeroConflict.R
\name{removeZeroConflict}
\alias{removeZeroConflict}
\title{Remove Zero Conflicts}
\usage{
removeZeroConflict(data, value1, value2, observationFlag1, observationFlag2,
  methodFlag1, methodFlag2, missingObservationFlag = "M",
  missingMethodFlag = "u")
}
\arguments{
\item{data}{The data table object.}

\item{value1}{The column name of data corresponding to the first variable.}

\item{value2}{The column name of data corresponding to the second variable.}

\item{observationFlag1}{The column name of data containing the observation
flag for the first variable.}

\item{observationFlag2}{The column name of data containing the observation
flag for the second variable.}

\item{methodFlag1}{The column name of data containing the method flag for
the first variable.}

\item{methodFlag2}{The column name of data containing the method flag for
the second variable.}

\item{missingObservationFlag}{The flag (character value) which should be
placed in the observation flag columns to signify a missing value.}

\item{missingMethodFlag}{The flag (character value) which should be placed
in the method flag columns to signify a missing value.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified (some values are marked as missing if the have
conflicting zeroes).
}
\description{
The function examines two variables of a data.table object.  If one
variable is 0 and another is not, then both variables are marked as
missing.  This is useful for variables which should always be zero
concurrently, such as area harvested and production.
}

