% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/updateMissingFlags.R
\name{updateMissingFlags}
\alias{updateMissingFlags}
\title{Update Missing Flags}
\usage{
updateMissingFlags(data, value, flag, naFlag = "M")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{value}{The value of the observation.}

\item{flag}{The flag of the observation.}

\item{naFlag}{The value of the flag which denotes missing value.}
}
\description{
Some data from the statistical working system may have missing values for
observations but the observation flag may indicate that this value is not
missing.  This function updates those flags by assigning them a missing
flag ("M") in the case where that observation is missing.  This function is
similar to remove0M, which assigns NA's to values where the observation
flag is "M" (sort of the opposite of this function).
}

